backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["shcats <shcats@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.88.0"
uvicorn = "^0.20.0"
sqlalchemy = "^1.4.45"
passlib = "^1.7.4"
bcrypt = "^4.0.1"
python-multipart = "^0.0.5"
aiofiles = "^22.1.0"
python-jose = "^3.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
